Move external socket code to model instance
