+++ model_external_socket/build_castor.xml
@@ -93,7 +93,7 @@
 
     <target name="castor-src" depends="init">
         <antcall target="castor-execute">
-            <param name="schema" value="xmlSchemas/externalModelStochModelFactory.xsd"/>
+            <param name="schema" value="xmlSchemas/externalSocketModelFactory.xsd"/>
             <param name="package" value="org.openda.${module}.io.castorgenerated"/>
         </antcall>
     </target>
