+++ model_external_socket/java/model_external_socket.iml
@@ -63,5 +63,6 @@
       </library>
     </orderEntry>
     <orderEntry type="library" name="CoreResources" level="project" />
+    <orderEntry type="module" module-name="application" />
   </component>
 </module>
