+++ model_external_socket/java/src/org/openda/externalsocket/ExternalSocketModelFactoryConfigReader.java
@@ -1,12 +1,12 @@
 package org.openda.externalsocket;
 
 import org.openda.model_external_socket.io.castorgenerated.ExternalModelParameterComplexType;
-import org.openda.model_external_socket.io.castorgenerated.ExternalModelStochModelFactoryConfigXML;
+import org.openda.model_external_socket.io.castorgenerated.ExternalSocketModelFactoryConfigXML;
 import org.openda.utils.io.CastorUtils;
 
 import java.io.File;
 
-public class ExternalModelStochModelFactoryConfigReader {
+public class ExternalSocketModelFactoryConfigReader {
 
 	private final int portNumber;
 	private final double[] values;
@@ -14,8 +14,8 @@ public class ExternalModelStochModelFactoryConfigReader {
 	private final double[] lowerBounds;
 	private final double[] upperBounds;
 
-	public ExternalModelStochModelFactoryConfigReader(File configFile) {
-		ExternalModelStochModelFactoryConfigXML castor = (ExternalModelStochModelFactoryConfigXML) CastorUtils.parse(configFile,ExternalModelStochModelFactoryConfigXML.class);
+	public ExternalSocketModelFactoryConfigReader(File configFile) {
+		ExternalSocketModelFactoryConfigXML castor = (ExternalSocketModelFactoryConfigXML) CastorUtils.parse(configFile, ExternalSocketModelFactoryConfigXML.class);
 		portNumber = castor.getPortNumber();
 		int parameterCount = castor.getParameterCount();
 		values = new double[parameterCount];
@@ -24,7 +24,7 @@ public class ExternalModelStochModelFactoryConfigReader {
 		upperBounds = new double[parameterCount];
 		for (int i = 0; i < parameterCount; i++) {
 			ExternalModelParameterComplexType parameter = castor.getParameter(i);
-			values[i] = parameter.getValue();
+			values[i] = parameter.getStartValue();
 			stdDev[i] = parameter.getStdDev();
 			lowerBounds[i] = parameter.hasLowerBound() ? parameter.getLowerBound() : Double.NaN;
 			upperBounds[i] = parameter.hasUpperBound() ? parameter.getUpperBound() : Double.NaN;
