--- model_external_socket/java/test/org/openda/externalsocket/testData/ExternalModelStochModelFactory.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<externalModelStochModelFactory xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.openda.org" xsi:schemaLocation="http://www.openda.org http://schemas.openda.org/externalModelStochModelFactory.xsd">
-  <portNumber>8124</portNumber>
-  <parameter value="1.0" stdDev="0.1" lowerBound="0.0" upperBound="100.0"/>
-  <parameter value="2.0" stdDev="0.2" lowerBound="1.0" upperBound="200.0"/>
-  <parameter value="3.0" stdDev="0.4" lowerBound="0.2" upperBound="300.0"/>
-</externalModelStochModelFactory>
