+++ model_external_socket/xmlSchemas/externalSocketModelFactory.xsd
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- edited with XMLSpy v2009 sp1 (http://www.altova.com) by ICT (Stichting Deltares) -->
 <xs:schema xmlns:openda="http://www.openda.org" xmlns:xs="http://www.w3.org/2001/XMLSchema" targetNamespace="http://www.openda.org" elementFormDefault="qualified" attributeFormDefault="unqualified">
-  <xs:element name="externalModelStochModelFactory" type="openda:ExternalModelStochModelFactoryConfigXML">
+  <xs:element name="externalSocketModelFactory" type="openda:ExternalSocketModelFactoryConfigXML">
     <xs:annotation>
       <xs:documentation>Configuration for an external model factory. It contains information on the socket port number and parameter initialisation values</xs:documentation>
     </xs:annotation>
   </xs:element>
-  <xs:complexType name="ExternalModelStochModelFactoryConfigXML">
+  <xs:complexType name="ExternalSocketModelFactoryConfigXML">
     <xs:sequence>
       <xs:element name="portNumber" type="xs:integer"/>
       <xs:element name="parameter" type="openda:ExternalModelParameterComplexType" maxOccurs="unbounded"/>
     </xs:sequence>
   </xs:complexType>
   <xs:complexType name="ExternalModelParameterComplexType">
-    <xs:attribute name="value" type="xs:double" use="required"/>
+    <xs:attribute name="startValue" type="xs:double" use="required"/>
     <xs:attribute name="stdDev" type="xs:double" use="required"/>
     <xs:attribute name="lowerBound" type="xs:double"/>
     <xs:attribute name="upperBound" type="xs:double"/>
